+++ apps/backoffice-v2/src/pages/Entity/components/EditableDetails/interfaces.ts
@@ -1,4 +1,11 @@
+import { AnyObject } from '@ballerine/ui';
 import { TDropdownOption } from './types';
+
+export type IEditableDetailsDocument = {
+  propertiesSchema: Record<string, unknown>;
+  id: string;
+  properties: Record<string, string>;
+};
 
 export interface IEditableDetails {
   data: Array<{
@@ -15,11 +22,10 @@ export interface IEditableDetails {
   }>;
   valueId: string;
   id: string;
-  documents: Array<{
-    propertiesSchema: Record<string, unknown>;
-    id: string;
-    properties: Record<string, string>;
-  }>;
+  documents: Array<IEditableDetailsDocument>;
   title: string;
   workflowId: string;
+  contextUpdateMethod?: 'base' | 'director';
+  onSubmit?: (document: AnyObject) => void;
+  config: Record<string, unknown>;
 }