+++ apps/backoffice-v2/src/pages/Entity/hooks/useTasks/selectors/selectWorkflowDocuments.ts
@@ -1,0 +1,5 @@
+import { AnyObject } from '@ballerine/ui';
+
+export const selectWorkflowDocuments = (workflow: unknown): AnyObject[] =>
+  //@ts-ignore
+  (workflow?.context?.documents as AnyObject[]) || ([] as AnyObject[]);