+++ apps/kyb-app/src/components/layouts/AppShell/Navigation.tsx
@@ -1,10 +1,11 @@
-import { useStateManagerContext } from '@/components/organisms/DynamicUI/StateManager/components/StateProvider';
-import { useDynamicUIContext } from '@/components/organisms/DynamicUI/hooks/useDynamicUIContext';
-import clsx from 'clsx';
 import { ArrowLeft } from 'lucide-react';
 import { useCallback } from 'react';
-import { usePageResolverContext } from '@/components/organisms/DynamicUI/PageResolver/hooks/usePageResolverContext';
 import { useTranslation } from 'react-i18next';
+
+import { useStateManagerContext } from '@/components/organisms/DynamicUI/StateManager/components/StateProvider';
+import { useDynamicUIContext } from '@/components/organisms/DynamicUI/hooks/useDynamicUIContext';
+import { usePageResolverContext } from '@/components/organisms/DynamicUI/PageResolver/hooks/usePageResolverContext';
+import { ctw } from '@ballerine/ui';
 
 export const Navigation = () => {
   const { state } = useDynamicUIContext();
@@ -21,7 +22,7 @@ export const Navigation = () => {
 
   return (
     <button
-      className={clsx('cursor-pointer select-none ', {
+      className={ctw('cursor-pointer select-none ', {
         'pointer-events-none opacity-50': isDisabled,
       })}
       aria-disabled={isDisabled}