+++ apps/kyb-app/src/components/organisms/UIRenderer/elements/JSONForm/components/DocumentField/DocumentField.tsx
@@ -1,32 +1,41 @@
 import { useEventEmitterLogic } from '@/components/organisms/DynamicUI/StateManager/components/ActionsHandler';
 import { useStateManagerContext } from '@/components/organisms/DynamicUI/StateManager/components/StateProvider';
+import { useDynamicUIContext } from '@/components/organisms/DynamicUI/hooks/useDynamicUIContext';
 import { useUIElementToolsLogic } from '@/components/organisms/DynamicUI/hooks/useUIStateLogic/hooks/useUIElementsStateLogic/hooks/useUIElementToolsLogic';
 import { ErrorField } from '@/components/organisms/DynamicUI/rule-engines';
-import { getDocumentFileIdPath } from '@/components/organisms/UIRenderer/elements/JSONForm/components/DocumentField/helpers/getDocumentFileIdPath';
+import { DocumentValueDestinationParser } from '@/components/organisms/UIRenderer/elements/JSONForm/components/DocumentField/helpers/document-value-destination-parser';
+import { serializeDocumentId } from '@/components/organisms/UIRenderer/elements/JSONForm/components/DocumentField/helpers/serialize-document-id';
+import { FileUploaderField } from '@/components/organisms/UIRenderer/elements/JSONForm/components/FileUploaderField';
+import { useFileRepository } from '@/components/organisms/UIRenderer/elements/JSONForm/components/FileUploaderField/hooks/useFileRepository';
+import { UploadFileFn } from '@/components/organisms/UIRenderer/elements/JSONForm/components/FileUploaderField/hooks/useFileUploading/types';
 import { useUIElementErrors } from '@/components/organisms/UIRenderer/hooks/useUIElementErrors/useUIElementErrors';
 import { useUIElementState } from '@/components/organisms/UIRenderer/hooks/useUIElementState';
 import { Document, UIElement } from '@/domains/collection-flow';
 import { fetchFile, uploadFile } from '@/domains/storage/storage.api';
 import { collectionFlowFileStorage } from '@/pages/CollectionFlow/collection-flow.file-storage';
-import { AnyObject, ErrorsList, FileInputAdapter, RJSFInputProps } from '@ballerine/ui';
+import { findDocumentSchemaByTypeAndCategory } from '@ballerine/common';
+import { AnyObject, ErrorsList, RJSFInputProps } from '@ballerine/ui';
 import { HTTPError } from 'ky';
 import get from 'lodash/get';
 import set from 'lodash/set';
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useCallback, useLayoutEffect, useMemo, useState } from 'react';
 
-interface DocumentFieldParams {
+export interface DocumentFieldParams {
   documentData: Partial<Document>;
 }
 
 export const DocumentField = (
-  props: RJSFInputProps & { definition: UIElement<DocumentFieldParams> },
+  props: RJSFInputProps<string> & { definition: UIElement<DocumentFieldParams> } & {
+    inputIndex: number | null;
+  },
 ) => {
-  const { definition, ...restProps } = props;
+  const { state } = useDynamicUIContext();
+  //@ts-ignore
+  const { definition, formData, inputIndex, onBlur, ...restProps } = props;
   const { stateApi } = useStateManagerContext();
   const { payload } = useStateManagerContext();
   const [fieldError, setFieldError] = useState<ErrorField | null>(null);
   const { options } = definition;
-  const [fileItem, setFile] = useState<File | null>(null);
 
   const { toggleElementLoading } = useUIElementToolsLogic(definition.name);
   const { state: elementState } = useUIElementState(definition);
@@ -34,84 +43,89 @@ export const DocumentField = (
   const documentDefinition = useMemo(
     () => ({
       ...definition,
-      valueDestination: `document-error-${definition.options.documentData.id}`,
+      valueDestination: `document-error-${serializeDocumentId(
+        //@ts-ignore
+        definition.options.documentData.id,
+        inputIndex,
+      )}`,
     }),
-    [definition],
+    [definition, inputIndex],
   );
 
   const sendEvent = useEventEmitterLogic(definition);
 
-  const { validationErrors, warnings } = useUIElementErrors(documentDefinition);
+  const getErrorKey = useCallback(() => documentDefinition.valueDestination, [documentDefinition]);
+  const { validationErrors, warnings } = useUIElementErrors(documentDefinition, getErrorKey);
   const { isTouched } = elementState;
 
   const fileId = useMemo(() => {
     if (!Array.isArray(payload.documents)) return null;
 
-    const document = payload.documents.find((document: Document) => {
-      return document?.id === definition.options.documentData.id;
-    }) as Document;
+    //@ts-ignore
+    const parser = new DocumentValueDestinationParser(definition.valueDestination);
+    const documentsPath = parser.extractRootPath();
+    const documentPagePath = parser.extractPagePath();
+    //@ts-ignore
+    const documents = (get(payload, documentsPath) as Document[]) || [];
 
-    const fileIdPath = getDocumentFileIdPath(definition);
+    const document = documents.find((document: Document) => {
+      //@ts-ignore
+      return document?.id === serializeDocumentId(definition.options.documentData.id, inputIndex);
+    });
 
-    // @ts-ignore
-    const fileId = get(document, fileIdPath) as string | null;
+    //@ts-ignore
+    const documentPage = get(document, documentPagePath) as Document['pages'][number];
+    const fileIdPath = parser.extractFileIdPath();
+
+    //@ts-ignore
+    const fileId = get(documentPage, fileIdPath) as string | null;
 
     return fileId;
-  }, [payload.documents, definition]);
+  }, [payload, definition, inputIndex]);
+  //@ts-ignore
+  const { file } = useFileRepository(collectionFlowFileStorage, fileId);
 
-  useEffect(() => {
+  useLayoutEffect(() => {
     if (!fileId) return;
 
     const persistedFile = collectionFlowFileStorage.getFileById(fileId);
 
-    if (persistedFile) {
-      setFile(persistedFile);
-    } else {
-      void fetchFile(fileId).then(file => {
-        // @ts-ignore
-        const createdFile = new File([''], file.fileNameInBucket || file.fileNameOnDisk, {
-          type: 'text/plain',
-        });
+    if (persistedFile) return;
 
-        collectionFlowFileStorage.registerFile(fileId, createdFile);
-        setFile(createdFile);
+    void fetchFile(fileId).then(file => {
+      const createdFile = new File([''], file.fileNameInBucket || file.fileNameOnDisk || '', {
+        type: 'text/plain',
       });
-    }
+
+      collectionFlowFileStorage.registerFile(fileId, createdFile);
+    });
   }, [fileId, toggleElementLoading]);
 
-  const handleChange = useCallback(
+  //@ts-ignore
+  const fileUploader: UploadFileFn = useCallback(
     async (file: File) => {
-      toggleElementLoading();
+      //@ts-ignore
+      const parser = new DocumentValueDestinationParser(definition.valueDestination);
 
       const context = stateApi.getContext();
-      let document = (context.documents as Document[]).find(
-        document => document && document.id === definition.options.documentData.id,
+      //@ts-ignore
+      const documents = (get(context, parser.extractRootPath()) as Document[]) || [];
+      const document = documents.find(
+        document =>
+          //@ts-ignore
+          document && document.id === serializeDocumentId(options.documentData.id, inputIndex),
       );
-      if (!document) {
-        document = options.documentData as Document;
-        context.documents = [...(context.documents as Document[]), document];
-        set(context, 'documents', context.documents);
-      }
-
-      const fileIdPath = getDocumentFileIdPath(definition);
 
       try {
+        toggleElementLoading();
+        //@ts-ignore
         const uploadResult = await uploadFile({ file });
-
-        // @ts-ignore
-        set(document, fileIdPath, uploadResult.id);
-        // set(document, fileTypePath, file.type);
-        set(document, 'decision', {});
-
-        stateApi.setContext(context);
-
-        collectionFlowFileStorage.registerFile(uploadResult.id, file);
-        setFile(file);
         setFieldError(null);
-        sendEvent('onChange');
-      } catch (err) {
-        if (err instanceof HTTPError) {
-          const response = (await err.response.json()) as AnyObject;
+
+        return { fileId: uploadResult.id };
+      } catch (error) {
+        if (error instanceof HTTPError) {
+          const response = (await error.response.json()) as AnyObject;
           setFieldError({
             // @ts-ignore
             fieldId: document.id,
@@ -121,20 +135,84 @@ export const DocumentField = (
           return;
         }
 
-        throw err;
+        throw error;
       } finally {
         toggleElementLoading();
       }
     },
-    [stateApi, options, definition, toggleElementLoading, sendEvent],
+    [stateApi, options, inputIndex, definition.valueDestination, toggleElementLoading],
+  );
+
+  const handleChange = useCallback(
+    (fileId: string) => {
+      //@ts-ignore
+      const destinationParser = new DocumentValueDestinationParser(definition.valueDestination);
+      const pathToDocumentsList = destinationParser.extractRootPath();
+      const pathToPage = destinationParser.extractPagePath();
+      const pathToFileId = destinationParser.extractFileIdPath();
+
+      const context = stateApi.getContext();
+      //@ts-ignore
+      const documents = (get(context, pathToDocumentsList) as Document[]) || [];
+
+      let document = documents.find(
+        document =>
+          document?.id ===
+          serializeDocumentId(definition.options?.documentData?.id || '', Number(inputIndex)),
+      );
+
+      if (!document) {
+        document = {
+          ...options.documentData,
+          //@ts-ignore
+          id: serializeDocumentId(options.documentData.id, inputIndex),
+          propertiesSchema:
+            findDocumentSchemaByTypeAndCategory(
+              //@ts-ignore
+              options.documentData.type,
+              options.documentData.category,
+            )?.propertiesSchema || undefined,
+        } as Document;
+        documents.push(document);
+        //@ts-ignore
+        set(context, pathToDocumentsList, documents);
+      }
+
+      //@ts-ignore
+      const documentPage =
+        //@ts-ignore
+        (get(document, pathToPage) as Document['pages'][number]) ||
+        //@ts-ignore
+        ({} as Document['pages'][number]);
+
+      // Assigning file properties
+      //@ts-ignore
+      set(documentPage, pathToFileId, fileId);
+      set(documentPage, 'fileName', file?.name);
+      set(documentPage, 'type', file?.type);
+
+      //@ts-ignore
+      set(document, pathToPage, documentPage);
+      set(document, 'decision', {});
+
+      stateApi.setContext(context);
+
+      sendEvent('onChange');
+    },
+    [stateApi, options, definition, inputIndex, file, sendEvent],
   );
 
   return (
     <div className="flex flex-col gap-2">
-      <FileInputAdapter
-        {...restProps}
-        disabled={elementState.isLoading || restProps.disabled}
-        formData={fileItem}
+      <FileUploaderField
+        uploadFile={fileUploader}
+        disabled={
+          //@ts-ignore
+          state.isRevision && warnings.length ? false : elementState.isLoading || restProps.disabled
+        }
+        fileId={fileId}
+        fileRepository={collectionFlowFileStorage}
+        onBlur={onBlur as () => void}
         onChange={handleChange}
       />
       {warnings.length ? <ErrorsList errors={warnings.map(err => err.message)} /> : null}