+++ apps/kyb-app/src/components/organisms/UIRenderer/elements/StepperUI/StepperUI.tsx
@@ -1,88 +1,106 @@
-import { Step } from '@/common/hooks/useStepper';
 import { Stepper } from '@/components/atoms/Stepper';
-import { Item } from '@/components/atoms/Stepper/Item';
 import { VerticalLayout } from '@/components/atoms/Stepper/layouts/Vertical';
 import { usePageResolverContext } from '@/components/organisms/DynamicUI/PageResolver/hooks/usePageResolverContext';
 import { useStateManagerContext } from '@/components/organisms/DynamicUI/StateManager/components/StateProvider';
 import { useDynamicUIContext } from '@/components/organisms/DynamicUI/hooks/useDynamicUIContext';
-import { useMemo, useRef } from 'react';
+import { useMemo, useState } from 'react';
 import { usePageContext } from '@/components/organisms/DynamicUI/Page';
 import { UIPage } from '@/domains/collection-flow';
-import { UIElementState } from '@/components/organisms/DynamicUI/hooks/useUIStateLogic/hooks/useUIElementsStateLogic/types';
 import { ErrorField } from '@/components/organisms/DynamicUI/rule-engines';
+import { CollectionFlowContext } from '@/domains/collection-flow/types/flow-context.types';
+import { isPageCompleted } from '@/helpers/prepareInitialUIState';
+import { UIElementState } from '@/components/organisms/DynamicUI/hooks/useUIStateLogic/hooks/useUIElementsStateLogic/types';
+import {
+  BreadcrumbItemInput,
+  Breadcrumbs,
+} from '@/components/atoms/Stepper/components/atoms/Breadcrumbs';
+import { ctw } from '@ballerine/ui';
 
 export const StepperUI = () => {
   const { state: uiState } = useDynamicUIContext();
   const { pages, currentPage } = usePageResolverContext();
-  const { state } = useStateManagerContext();
+  const { payload } = useStateManagerContext();
   const { pageErrors } = usePageContext();
 
-  const initialPageNumber = useRef(currentPage?.number);
-
   const computeStepStatus = ({
-    uiElement,
-    page,
     pageError,
-    currentPage,
+    page,
+    context,
+    uiElementState,
   }: {
-    uiElement: UIElementState;
     page: UIPage;
+    uiElementState: UIElementState;
     pageError: Record<string, ErrorField>;
     currentPage: UIPage;
+    context: CollectionFlowContext;
   }) => {
-    if (!!Object.keys(pageError).length && currentPage.number === page.number) return 'warning';
-    if (
-      uiElement?.isCompleted ||
-      page.number <=
-        // @ts-ignore
-        initialPageNumber?.current
-    )
-      return 'completed';
+    if (Object.values(pageError || {}).some(error => error.type === 'warning')) return 'warning';
+
+    if (isPageCompleted(page, context) || uiElementState?.isCompleted) return 'completed';
 
     return 'idle';
   };
 
-  const steps: Step[] = useMemo(() => {
+  const [initialContext] = useState(() => structuredClone(payload));
+
+  const steps: BreadcrumbItemInput[] = useMemo(() => {
     return pages.map(page => {
       const stepStatus = computeStepStatus({
         // @ts-ignore
-        uiElement: uiState.elements[page.stateName],
-        page,
+        uiElementState: uiState.elements[page.stateName],
         // @ts-ignore
         pageError: pageErrors?.[page.stateName],
-        // @ts-ignore
-        currentPage,
+        page,
+        context: initialContext,
+        currentPage: currentPage as UIPage,
       });
 
-      const step: Step = {
-        index: page.number - 1,
+      const step: BreadcrumbItemInput = {
+        id: page.stateName,
         label: page.name,
-        dataAlias: page.stateName,
-        status: stepStatus,
-        meta: { status: stepStatus },
-        // @ts-ignore
-        isCurrent: currentPage?.number === page.number,
+        state: stepStatus,
       };
 
       return step;
     });
-  }, [pages, uiState.elements, pageErrors, currentPage]);
+  }, [pages, uiState, pageErrors, initialContext, currentPage]);
+
+  const activeStep = useMemo(() => {
+    const activeStep = steps.find(step => step.id === currentPage?.stateName);
+    if (!activeStep) return null;
+
+    return activeStep;
+  }, [steps, currentPage]);
 
   return (
     <Stepper>
-      <VerticalLayout>
-        {steps.map(step => {
-          return step.hidden ? null : (
-            <Item
-              key={`step-${step.index}`}
-              active={state === step.dataAlias}
-              label={step.label}
-              // @ts-ignore
-              status={step.meta?.status}
-            />
+      <Breadcrumbs items={steps} active={activeStep}>
+        {(items, theme) => {
+          return (
+            <VerticalLayout>
+              {items.map(itemProps => {
+                return (
+                  <div
+                    className={ctw('last:bg- flex flex-row items-center gap-4 first:bg-white')}
+                    key={itemProps.id}
+                  >
+                    <Breadcrumbs.Item
+                      active={itemProps.active}
+                      state={itemProps.state}
+                      theme={theme}
+                    />
+                    <Breadcrumbs.Label
+                      active={itemProps.active}
+                      text={itemProps.label}
+                      state={itemProps.state}
+                    />
+                  </div>
+                );
+              })}
+            </VerticalLayout>
           );
-        })}
-      </VerticalLayout>
+        }}
+      </Breadcrumbs>
     </Stepper>
   );
 };