+++ apps/kyb-app/src/pages/CollectionFlow/CollectionFlow.tsx
@@ -56,10 +56,10 @@ export const CollectionFlow = withSessionProtected(() => {
   const elements = schema?.uiSchema?.elements;
   const definition = schema?.definition.definition;
 
-  const pageErrors = usePageErrors(
-    context ?? {},
-    // @ts-ignore
-    elements,
+  const pageErrors = usePageErrors(context ?? {}, elements || []);
+  const isRevision = useMemo(
+    () => pageErrors.some(error => error.errors?.some(error => error.type === 'warning')),
+    [pageErrors],
   );
 
   const filteredNonEmptyErrors = pageErrors?.filter(pageError => !!pageError.errors.length);
@@ -83,8 +83,8 @@ export const CollectionFlow = withSessionProtected(() => {
   }, [context, elements, filteredNonEmptyErrors]);
 
   const initialUIState = useMemo(() => {
-    return prepareInitialUIState(elements || [], context || {});
-  }, [elements, context]);
+    return prepareInitialUIState(elements || [], context || {}, isRevision);
+  }, [elements, context, isRevision]);
 
   if (initialContext?.flowConfig?.appState === 'approved') return <Approved />;
   if (initialContext?.flowConfig?.appState == 'rejected') return <Rejected />;