+++ apps/kyb-app/src/pages/CollectionFlow/collection-flow.file-storage.ts
@@ -1,3 +1,3 @@
-import { FileStorage } from '@/utils/file-storage';
+import { FileRepository } from '@/utils/file-repository';
 
-export const collectionFlowFileStorage = new FileStorage();
+export const collectionFlowFileStorage = new FileRepository();