+++ packages/common/src/schemas/documents/workflow/documents/schemas/index.ts
@@ -1,10 +1,10 @@
-import { getGhanaDocuments } from './GH';
-import { TDocument } from '../types';
 import { countryCodes } from '@/countries';
 import { DefaultContextSchema } from '@/schemas';
-import { getCanadaDocuments } from './CA';
-import { getUniversalDocuments } from './ZZ';
 import { getUgandaDocuments } from '@/schemas/documents/workflow/documents/schemas/UG';
+import { TDocument } from '../types';
+import { getCanadaDocuments } from './CA';
+import { getGhanaDocuments } from './GH';
+import { getUniversalDocuments } from './ZZ';
 
 const documentIdsByCountry: Partial<Record<(typeof countryCodes)[number], () => TDocument[]>> = {
   GH: getGhanaDocuments,
@@ -15,6 +15,15 @@ const documentIdsByCountry: Partial<Record<(typeof countryCodes)[number], () => 
 
 export const getDocumentsByCountry = (countryCode: (typeof countryCodes)[number]): TDocument[] => {
   return documentIdsByCountry[countryCode]?.() || [];
+};
+
+export const getDocumentSchemaByCountry = (
+  countryCode: (typeof countryCodes)[number],
+  documentsSchema: TDocument[] | undefined,
+): TDocument[] => {
+  return (
+    documentsSchema?.filter(documentSchema => documentSchema.issuer.country === countryCode) || []
+  );
 };
 
 export const getDocumentId = (
@@ -30,4 +39,22 @@ export const getDocumentId = (
   }
 
   return id.toLowerCase();
+};
+
+export const findDocumentSchemaByTypeAndCategory = (
+  type: string,
+  category: string,
+): TDocument | null => {
+  const documentGetters = Object.values(documentIdsByCountry);
+
+  for (const getDocuments of documentGetters) {
+    const documents = getDocuments();
+    const document = documents.find(doc => doc.type === type && doc.category === category);
+
+    if (!document) continue;
+
+    return document;
+  }
+
+  return null;
 };