+++ packages/common/src/utils/replace-null-with-undefined/replace-null-with-undefined.ts
@@ -1,0 +1,16 @@
+export const replaceNullsWithUndefined = (obj: unknown) => {
+  if (obj === null) {
+    return undefined;
+  }
+
+  if (typeof obj !== 'object' || Array.isArray(obj)) {
+    return obj;
+  }
+
+  const objWithoutNulls = Object.entries(obj).reduce((acc, [key, value]) => {
+    acc[key] = replaceNullsWithUndefined(value);
+    return acc;
+  }, {} as Record<PropertyKey, unknown>);
+
+  return objWithoutNulls;
+};