+++ packages/ui/src/components/molecules/inputs/MultiSelect/MultiSelect.stories.tsx
@@ -39,3 +39,33 @@ const DefaultComponent = () => {
 export const Default = {
   render: DefaultComponent,
 };
+
+const DisabledComponent = () => {
+  const [value, setValue] = useState<Array<MultiSelectValue>>([]);
+
+  const renderSelected: MultiSelectSelectedItemRenderer = useCallback((params, option) => {
+    return (
+      <Chip key={option.value} className="h-6">
+        <Chip.Label text={option.title} variant="secondary" />
+        <Chip.UnselectButton
+          {...params.unselectButtonProps}
+          icon={<X className="hover:text-muted-foreground h-3 w-3 text-white" />}
+        />
+      </Chip>
+    );
+  }, []);
+
+  return (
+    <MultiSelect
+      value={value}
+      options={options}
+      disabled
+      onChange={setValue}
+      renderSelected={renderSelected}
+    />
+  );
+};
+
+export const Disabled = {
+  render: DisabledComponent,
+};