+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @ballerine/workflow-node-sdk
 
+## 0.5.28
+
+### Patch Changes
+
+- Updated dependencies [85d24981]
+  - @ballerine/workflow-core@0.5.28
+
+## 0.5.27
+
+### Patch Changes
+
+- Updated dependencies [85d24981]
+  - @ballerine/workflow-core@0.5.27
+
+## 0.5.26
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.5.26
+
+## 0.5.25
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.5.25
+
+## 0.5.24
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.5.24
+
 ## 0.5.23
 
 ### Patch Changes
@@ -28,6 +60,7 @@
 
 ### Patch Changes
 
+- @ballerine/workflow-core@0.5.20
 - Change document schemas
 - Updated dependencies
   - @ballerine/workflow-core@0.5.20